core: improve BioPax parser by reusing Xref, Ontology and Cellular Location nodes

+++ bionetdb-core/src/main/java/org/opencb/bionetdb/core/network/Node.java
@@ -97,7 +97,7 @@ public class Node {
         }
     }
 
-    public Node(int uid) {
+    public Node(long uid) {
         tags = new ArrayList<>();
         attributes = new ObjectMap();
